update product list
