+++ ECommerce.WebApp/client/src/_pages/product-list/_components/index.ts
@@ -1,0 +1,4 @@
+export { default as ProlCategory } from "./prol-category";
+export { default as ProlCategoryMb } from "./mobile/prol-category-mb";
+export { default as ProlFilter } from "./prol-filter";
+export { default as ProlPagination } from './prol-pagination';
